+++ packages/twenty-server/src/modules/messaging/message-import-manager/drivers/gmail/messaging-gmail-driver.module.ts
@@ -9,24 +9,15 @@ import { WorkspaceDataSourceModule } from 'src/engine/workspace-datasource/works
 import { BlocklistWorkspaceEntity } from 'src/modules/blocklist/standard-objects/blocklist.workspace-entity';
 import { EmailAliasManagerModule } from 'src/modules/connected-account/email-alias-manager/email-alias-manager.module';
 import { OAuth2ClientManagerModule } from 'src/modules/connected-account/oauth2-client-manager/oauth2-client-manager.module';
-import { RefreshAccessTokenManagerModule } from 'src/modules/connected-account/refresh-access-token-manager/refresh-access-token-manager.module';
 import { ConnectedAccountWorkspaceEntity } from 'src/modules/connected-account/standard-objects/connected-account.workspace-entity';
 import { MessagingCommonModule } from 'src/modules/messaging/common/messaging-common.module';
 import { MessageChannelMessageAssociationWorkspaceEntity } from 'src/modules/messaging/common/standard-objects/message-channel-message-association.workspace-entity';
 import { MessageChannelWorkspaceEntity } from 'src/modules/messaging/common/standard-objects/message-channel.workspace-entity';
 import { MessagingGmailClientProvider } from 'src/modules/messaging/message-import-manager/drivers/gmail/providers/messaging-gmail-client.provider';
-import { MessagingGmailFetchMessagesByBatchesService } from 'src/modules/messaging/message-import-manager/drivers/gmail/services/messaging-gmail-fetch-messages-by-batches.service';
-import { MessagingGmailFetchMessageIdsToExcludeService } from 'src/modules/messaging/message-import-manager/drivers/gmail/services/messaging-gmail-fetch-messages-ids-to-exclude.service';
-import { MessagingGmailFullMessageListFetchService } from 'src/modules/messaging/message-import-manager/drivers/gmail/services/messaging-gmail-full-message-list-fetch.service';
-import { MessagingGmailHistoryService } from 'src/modules/messaging/message-import-manager/drivers/gmail/services/messaging-gmail-history.service';
-import { MessagingGmailMessagesImportService } from 'src/modules/messaging/message-import-manager/drivers/gmail/services/messaging-gmail-messages-import.service';
-import { MessagingGmailPartialMessageListFetchService } from 'src/modules/messaging/message-import-manager/drivers/gmail/services/messaging-gmail-partial-message-list-fetch.service';
-import { MessagingSaveMessagesAndEnqueueContactCreationService } from 'src/modules/messaging/message-import-manager/drivers/gmail/services/messaging-save-messages-and-enqueue-contact-creation.service';
 import { MessageParticipantManagerModule } from 'src/modules/messaging/message-participant-manager/message-participant-manager.module';
 
 @Module({
   imports: [
-    RefreshAccessTokenManagerModule,
     EnvironmentModule,
     ObjectMetadataRepositoryModule.forFeature([
       ConnectedAccountWorkspaceEntity,
@@ -42,24 +33,7 @@ import { MessageParticipantManagerModule } from 'src/modules/messaging/message-p
     WorkspaceDataSourceModule,
     MessageParticipantManagerModule,
   ],
-  providers: [
-    MessagingGmailClientProvider,
-    MessagingGmailHistoryService,
-    MessagingGmailFetchMessagesByBatchesService,
-    MessagingGmailPartialMessageListFetchService,
-    MessagingGmailFullMessageListFetchService,
-    MessagingGmailMessagesImportService,
-    MessagingGmailFetchMessageIdsToExcludeService,
-    MessagingSaveMessagesAndEnqueueContactCreationService,
-  ],
-  exports: [
-    MessagingGmailClientProvider,
-    MessagingGmailHistoryService,
-    MessagingGmailFetchMessagesByBatchesService,
-    MessagingGmailPartialMessageListFetchService,
-    MessagingGmailFullMessageListFetchService,
-    MessagingGmailMessagesImportService,
-    MessagingGmailFetchMessageIdsToExcludeService,
-  ],
+  providers: [MessagingGmailClientProvider],
+  exports: [MessagingGmailClientProvider],
 })
 export class MessagingGmailDriverModule {}
