+++ packages/twenty-server/src/modules/messaging/message-import-manager/messaging-import-manager.module.ts
@@ -1,9 +1,14 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
+import { BillingModule } from 'src/engine/core-modules/billing/billing.module';
+import { FeatureFlagModule } from 'src/engine/core-modules/feature-flag/feature-flag.module';
 import { Workspace } from 'src/engine/core-modules/workspace/workspace.entity';
 import { DataSourceEntity } from 'src/engine/metadata-modules/data-source/data-source.entity';
 import { TwentyORMModule } from 'src/engine/twenty-orm/twenty-orm.module';
+import { WorkspaceDataSourceModule } from 'src/engine/workspace-datasource/workspace-datasource.module';
+import { EmailAliasManagerModule } from 'src/modules/connected-account/email-alias-manager/email-alias-manager.module';
+import { RefreshAccessTokenManagerModule } from 'src/modules/connected-account/refresh-access-token-manager/refresh-access-token-manager.module';
 import { MessagingCommonModule } from 'src/modules/messaging/common/messaging-common.module';
 import { MessageChannelWorkspaceEntity } from 'src/modules/messaging/common/standard-objects/message-channel.workspace-entity';
 import { MessagingSingleMessageImportCommand } from 'src/modules/messaging/message-import-manager/commands/messaging-single-message-import.command';
@@ -20,11 +25,21 @@ import { MessagingMessageListFetchJob } from 'src/modules/messaging/message-impo
 import { MessagingMessagesImportJob } from 'src/modules/messaging/message-import-manager/jobs/messaging-messages-import.job';
 import { MessagingOngoingStaleJob } from 'src/modules/messaging/message-import-manager/jobs/messaging-ongoing-stale.job';
 import { MessagingMessageImportManagerMessageChannelListener } from 'src/modules/messaging/message-import-manager/listeners/messaging-import-manager-message-channel.listener';
-import { BillingModule } from 'src/engine/core-modules/billing/billing.module';
-import { WorkspaceDataSourceModule } from 'src/engine/workspace-datasource/workspace-datasource.module';
+import { MessagingErrorHandlingService } from 'src/modules/messaging/message-import-manager/services/messaging-error-handling.service';
+import { MessagingGmailFetchMessagesByBatchesService } from 'src/modules/messaging/message-import-manager/services/messaging-gmail-fetch-messages-by-batches.service';
+import { MessagingGmailFetchMessageIdsToExcludeService } from 'src/modules/messaging/message-import-manager/services/messaging-gmail-fetch-messages-ids-to-exclude.service';
+import { MessagingGmailFullMessageListFetchService } from 'src/modules/messaging/message-import-manager/services/messaging-gmail-full-message-list-fetch.service';
+import { MessagingGmailHistoryService } from 'src/modules/messaging/message-import-manager/services/messaging-gmail-history.service';
+import { MessagingGmailMessagesImportService } from 'src/modules/messaging/message-import-manager/services/messaging-gmail-messages-import.service';
+import { MessagingGmailPartialMessageListFetchService } from 'src/modules/messaging/message-import-manager/services/messaging-gmail-partial-message-list-fetch.service';
+import { MessagingMessageThreadService } from 'src/modules/messaging/message-import-manager/services/messaging-message-thread.service';
+import { MessagingMessageService } from 'src/modules/messaging/message-import-manager/services/messaging-message.service';
+import { MessagingSaveMessagesAndEnqueueContactCreationService } from 'src/modules/messaging/message-import-manager/services/messaging-save-messages-and-enqueue-contact-creation.service';
+import { MessageParticipantManagerModule } from 'src/modules/messaging/message-participant-manager/message-participant-manager.module';
 
 @Module({
   imports: [
+    RefreshAccessTokenManagerModule,
     WorkspaceDataSourceModule,
     MessagingGmailDriverModule,
     MessagingCommonModule,
@@ -32,6 +47,9 @@ import { WorkspaceDataSourceModule } from 'src/engine/workspace-datasource/works
     TypeOrmModule.forFeature([DataSourceEntity], 'metadata'),
     TwentyORMModule.forFeature([MessageChannelWorkspaceEntity]),
     BillingModule,
+    EmailAliasManagerModule,
+    FeatureFlagModule,
+    MessageParticipantManagerModule,
   ],
   providers: [
     MessagingMessageListFetchCronCommand,
@@ -47,6 +65,16 @@ import { WorkspaceDataSourceModule } from 'src/engine/workspace-datasource/works
     MessagingAddSingleMessageToCacheForImportJob,
     MessagingMessageImportManagerMessageChannelListener,
     MessagingCleanCacheJob,
+    MessagingMessageService,
+    MessagingMessageThreadService,
+    MessagingErrorHandlingService,
+    MessagingGmailHistoryService,
+    MessagingGmailFetchMessagesByBatchesService,
+    MessagingGmailPartialMessageListFetchService,
+    MessagingGmailFullMessageListFetchService,
+    MessagingGmailMessagesImportService,
+    MessagingGmailFetchMessageIdsToExcludeService,
+    MessagingSaveMessagesAndEnqueueContactCreationService,
   ],
   exports: [],
 })
