+++ packages/twenty-server/src/modules/messaging/message-import-manager/services/messaging-gmail-fetch-messages-ids-to-exclude.service.ts
@@ -3,8 +3,8 @@ import { Injectable } from '@nestjs/common';
 import { gmail_v1 } from 'googleapis';
 
 import { MESSAGING_GMAIL_EXCLUDED_CATEGORIES } from 'src/modules/messaging/message-import-manager/drivers/gmail/constants/messaging-gmail-excluded-categories';
-import { MessagingGmailHistoryService } from 'src/modules/messaging/message-import-manager/drivers/gmail/services/messaging-gmail-history.service';
 import { computeGmailCategoryLabelId } from 'src/modules/messaging/message-import-manager/drivers/gmail/utils/compute-gmail-category-label-id';
+import { MessagingGmailHistoryService } from 'src/modules/messaging/message-import-manager/services/messaging-gmail-history.service';
 import { assertNotNull } from 'src/utils/assert';
 
 @Injectable()
