+++ packages/twenty-server/src/modules/messaging/message-import-manager/services/messaging-gmail-history.service.ts
@@ -1,10 +1,10 @@
 import { Injectable } from '@nestjs/common';
 
-import { gmail_v1 } from 'googleapis';
 import { GaxiosResponse } from 'gaxios';
+import { gmail_v1 } from 'googleapis';
 
-import { GmailError } from 'src/modules/messaging/common/services/messaging-error-handling.service';
 import { MESSAGING_GMAIL_USERS_HISTORY_MAX_RESULT } from 'src/modules/messaging/message-import-manager/drivers/gmail/constants/messaging-gmail-users-history-max-result.constant';
+import { GmailError } from 'src/modules/messaging/message-import-manager/services/messaging-error-handling.service';
 
 @Injectable()
 export class MessagingGmailHistoryService {
